Move simplecov setup
