+++ spec/rails_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'simplecov_setup'
+require 'support/simplecov'
 require 'spec_helper'
 ENV['RAILS_ENV'] ||= 'test'
 require File.expand_path('../config/environment', __dir__)
