+++ spec/support/simplecov.rb
